refactor room management service
